Move shared rflink constants to separate module
