+++ homeassistant/components/rflink/binary_sensor.py
@@ -26,7 +26,8 @@ import homeassistant.helpers.event as evt
 from homeassistant.helpers.restore_state import RestoreEntity
 from homeassistant.helpers.typing import ConfigType, DiscoveryInfoType
 
-from . import CONF_ALIASES, RflinkDevice
+from . import RflinkDevice
+from .const import CONF_ALIASES
 
 CONF_OFF_DELAY = "off_delay"
 DEFAULT_FORCE_UPDATE = False
